+++ test/framework/GoogleRunnable.test.ts
@@ -24,6 +24,7 @@ describe(pathlib.basename(__filename), function () {
     {},
     {},
     1,
+    false,
     {},
     new DefaultSpawner(),
   );